# Add kmer heuristic for faster adapter matching