--- a/src/cutadapt/_kmer_finder.pyi
+++ b/src/cutadapt/_kmer_finder.pyi
@@ -1,9 +1,9 @@
-from typing import List, Tuple
+from typing import List, Optional, Tuple
 
 class KmerFinder:
     def __init__(
         self,
-        kmers_and_positions: List[Tuple[str, int, int]],
+        kmers_and_positions: List[Tuple[str, int, Optional[int]]],
         ref_wildcards: bool,
         query_wildcards: bool,
     ): ...