--- a/src/cutadapt/_kmer_finder.pyi
+++ b/src/cutadapt/_kmer_finder.pyi
@@ -4,7 +4,7 @@ class KmerFinder:
     def __init__(
         self,
         kmers_and_positions: List[Tuple[str, int, Optional[int]]],
-        ref_wildcards: bool,
-        query_wildcards: bool,
+        ref_wildcards: bool = False,
+        query_wildcards: bool = False,
     ): ...
     def kmers_present(self, __sequence: str) -> bool: ...