--- a/src/cutadapt/_kmer_finder.pyi
+++ b/src/cutadapt/_kmer_finder.pyi
@@ -3,7 +3,7 @@ from typing import List, Optional, Tuple
 class KmerFinder:
     def __init__(
         self,
-        kmers_and_positions: List[Tuple[int, Optional[int], List[str]]],
+        positions_and_kmers: List[Tuple[int, Optional[int], List[str]]],
         ref_wildcards: bool = False,
         query_wildcards: bool = False,
     ): ...