--- a/src/cutadapt/_kmer_finder.pyx
+++ b/src/cutadapt/_kmer_finder.pyx
@@ -15,21 +15,22 @@ cdef extern from "Python.h":
 ctypedef struct KmerEntry:
     size_t kmer_length
     size_t mask_offset
-    ssize_t search_offset
+    ssize_t search_start
+    ssize_t search_stop
 
 
 cdef class KmerFinder:
     """
     Find kmers in strings. To replace the following code:
 
-        kmers_and_offsets = [("AGA", -10), ("AGCATGA", 0)]
-        for kmer, offset in kmers_and_offsets:
-            sequence.find(kmer, offset)
+        kmers_and_positions = [("AGA", -10, None), ("AGCATGA", 0, None)]
+        for kmer, start, stop in kmers_and_positions:
+            sequence.find(kmer, start, stop)
 
     This has a lot of python overhead. The following code is equivalent:
 
-        kmers_and_offsets = [("AGA", -10), ("AGCATGA", 0)]
-        kmer_finder = KmerFinder(kmers_and_offsets)
+        kmers_and_positions = [("AGA", -10, None), ("AGCATGA", 0, None)]
+        kmer_finder = KmerFinder(kmers_and_positions)
         kmer_finder.kmers_present(sequence)
 
     This is more efficient as the kmers_present method can be applied to a lot
@@ -40,19 +41,19 @@ cdef class KmerFinder:
         KmerEntry *kmer_entries
         size_t *kmer_masks
         size_t number_of_kmers
-        object kmers_and_offsets
+        object kmers_and_positions
         readonly bint ref_wildcards
         readonly bint query_wildcards
 
-    def __cinit__(self, kmers_and_offsets, ref_wildcards=False, query_wildcards=False):
+    def __cinit__(self, kmers_and_positions, ref_wildcards=False, query_wildcards=False):
         self.kmer_masks = NULL
         self.kmer_entries = NULL
         self.number_of_kmers = 0
         self.ref_wildcards = ref_wildcards
         self.query_wildcards = query_wildcards
-        kmers = [kmer for kmer, _ in kmers_and_offsets]
+        kmers = [kmer for kmer, _ in kmers_and_positions]
         kmer_total_length = sum(len(kmer) for kmer in kmers)
-        number_of_entries = len(kmers_and_offsets)
+        number_of_entries = len(kmers_and_positions)
         self.kmer_entries = <KmerEntry *>PyMem_Malloc(number_of_entries * sizeof(KmerEntry))
         # for the kmers the NULL bytes also need space.
         self.kmer_masks = <size_t *>PyMem_Malloc(number_of_entries * sizeof(size_t) * ASCII_CHAR_COUNT)
@@ -60,12 +61,15 @@ cdef class KmerFinder:
         cdef size_t mask_offset = 0
         cdef char *kmer_ptr
         cdef Py_ssize_t kmer_length
-        for i, (kmer, offset) in enumerate(kmers_and_offsets):
+        for i, (kmer, start, stop) in enumerate(kmers_and_positions):
             if not PyUnicode_CheckExact(kmer):
                 raise TypeError(f"Kmer should be a string not {type(kmer)}")
             if not PyUnicode_IS_COMPACT_ASCII(kmer):
                 raise ValueError("Only ASCII strings are supported")
-            self.kmer_entries[i].search_offset  = offset
+            self.kmer_entries[i].search_start  = start
+            if stop is None:
+                stop = 0
+            self.kmer_intries[i].search_stop = stop
             self.kmer_entries[i].mask_offset = mask_offset
             kmer_length = PyUnicode_GET_LENGTH(kmer)
             self.kmer_entries[i].kmer_length = kmer_length
@@ -73,10 +77,10 @@ cdef class KmerFinder:
             populate_needle_mask(self.kmer_masks + mask_offset, kmer_ptr, kmer_length,
                                  self.ref_wildcards, self.query_wildcards)
             mask_offset += ASCII_CHAR_COUNT
-        self.kmers_and_offsets = kmers_and_offsets
+        self.kmers_and_positions = kmers_and_positions
 
     def __reduce__(self):
-        return KmerFinder, (self.kmers_and_offsets,)
+        return KmerFinder, (self.kmers_and_positions,)
 
     def kmers_present(self, str sequence):
         cdef:
@@ -84,7 +88,7 @@ cdef class KmerFinder:
             size_t i
             size_t kmer_offset
             size_t kmer_length
-            ssize_t search_offset
+            ssize_t start, stop
             size_t *mask_ptr
             char *search_ptr
             char *search_result
@@ -95,19 +99,28 @@ cdef class KmerFinder:
         cdef Py_ssize_t seq_length = PyUnicode_GET_LENGTH(sequence)
         for i in range(self.number_of_kmers):
             entry = self.kmer_entries[i]
-            search_offset = entry.search_offset
-            if search_offset < 0:
-                search_offset = seq_length + search_offset
-                if search_offset < 0:
-                    search_offset = 0
-            if search_offset > seq_length:
+            start = entry.search_start 
+            stop = entry.search_stop
+            if start < 0:
+                start = seq_length + start
+                if start < 0:
+                    start = 0
+            if start > seq_length:
                 continue
+            if stop < 0:
+                step = seq_length + stop
+                if stop <= 0:  # No need to search
+                    continue
+            if stop == 0:
+                stop = seq_length
+            search_length = stop - start
+            if search_length <= 0:
+                continue
+            search_ptr = seq + start
             kmer_length = entry.kmer_length
             mask_ptr = self.kmer_masks + entry.mask_offset
-            search_ptr = seq + search_offset
-            search_length = seq_length - (search_ptr - seq)
             search_result = shift_or_search(search_ptr, search_length,
-                                             mask_ptr, kmer_length)
+                                            mask_ptr, kmer_length)
             if search_result:
                 return True
         return False