--- a/src/cutadapt/_kmer_finder.pyx
+++ b/src/cutadapt/_kmer_finder.pyx
@@ -135,7 +135,7 @@ cdef void set_masks(size_t *needle_mask, size_t pos, char *chars):
         needle_mask[<uint8_t>chars[i]] &= ~(1UL << pos)
 
 cdef populate_needle_mask(size_t *needle_mask, char *needle, size_t needle_length,
-                          bint ref_wildcards, query_wildcards):
+                          bint ref_wildcards, bint query_wildcards):
     cdef size_t i
     cdef char c
     if needle_length > (sizeof(size_t) * 8 - 1):
@@ -207,7 +207,12 @@ cdef populate_needle_mask(size_t *needle_mask, char *needle, size_t needle_lengt
             if query_wildcards:  # Proper IUPAC matching
                 set_masks(needle_mask, i, "URYSWKMBDHVNuryswkmbdhvn")
             else:  # N matches literally everything except \00
-                set_masks(needle_mask, i, bytes(range(1, 128)))
+                set_masks(needle_mask, i,
+                          "\x01\x02\x03\x04\x05\x06\x07\x08\t\n\x0b\x0c\r\x0e"
+                          "\x0f\x10\x11\x12\x13\x14\x15\x16\x17\x18\x19\x1a"
+                          "\x1b\x1c\x1d\x1e\x1f !\"#$%&'()*+,-./0123456789:;<="
+                          ">?@ABCDEFGHIJKLMNOPQRSTUVWXYZ[\\]^_`abcdefghijklmno"
+                          "pqrstuvwxyz{|}~\x7f")
         else:
             needle_mask[<uint8_t>c] &= ~(1UL << i)