… src/cutadapt/_match_tables.py
… src/cutadapt/adapters.py
… src/cutadapt/kmer_heuristic.py
… tests/test_kmer_finder.py
… tests/test_kmer_heuristic.py
… tox.ini