Alignment is the most cost effective way for approximately matching an adapter to a read sequence at the best possible position. The cost of this is alignment is proportional to the size of the adapter (m) multiplied by the length of the sequence (n). O(mn) cost means that the alignment is the most costly part of cutadapt.

There is however a subproblem that is much easier to solve: determining if an adapter is at all present in a sequence. If x errors are allowed, we can subdivide the adapter in x+1 pieces. It follows that at least one of the pieces must be present in the sequence without error. Modern search algorithms operate at a speed of near O(n) even for multiple pattern matching. When a sequence is determined not to be present in a read, alignment can be skipped. This results in a much reduced time to completion.

This PR adds a kmer_heuristic.py file which provides an algorithm for cutting up the adapter in pieces at least one of which must occur. It has code to handle adapters that only partially overlap at the beginning of the end and can handle internal and non-internal adapters.

It also provides the shift-OR string searching algorithm. This algorithm uses a lookup table of bitmasks to determine if a word matches. The bitmasks can be created at the program initialization rather than at string searching time giving this algorithm an advantage over generic string searching functions such as CPython's find method or C stdlib's strstr. This means the seach algorithm runs at O(n) rather than O(m+n) since the initialization cost is not payed for each read. Another advantage of shift-OR algorithm is that matching IUPAC costs the same as direct matching. The KmerFinder class created in its own Cython file handles the creation and storage of the bitmasks, as well as the searching. The shift-OR algorithm was slightly altered to store multiple words in each bit matrix so all words can be foud in O(n) time while performing multiple pattern matching.

Preliminary benchmarks show that the heuristic runs 20 times faster than the alignment. This means that theoretically the cost of the heuristic is worth it unless the adapter content is above 95%. Since the heuristic exits early when one of the sequences is found, the tipping point might actually be higher than 95. 

EDIT: Updated to reflect PR content. This makes for a better merge commit message. Original message below.
EDIT2: Further updates including the O(n) nature of the updated algorithm.

<details><summary>Original message</summary>

Hi, just a draft PR to investigate this new possibility I informed you about. 
cProfile result before:
```
  5000000    1.486    0.000   24.305    0.000 adapters.py:749(match_to)
  5000000   22.613    0.000   22.613    0.000 {method 'locate' of 'cutadapt._align.Aligner' objects}
   559384    0.390    0.000    0.667    0.000 adapters.py:177(add_match)
```

After:
```
  5000000    1.413    0.000   11.415    0.000 adapters.py:761(match_to)
  5000000    5.639    0.000    5.639    0.000 {method 'kmers_present' of 'cutadapt._kmer_finder.KmerFinder' objects}
   797111    4.157    0.000    4.157    0.000 {method 'locate' of 'cutadapt._align.Aligner' objects}
   559384    0.380    0.000    0.655    0.000 adapters.py:179(add_match)
```
As you can see, the heuristic is not cheap it roughly requires 25% of the alignment resources. It is quite effective though in eliminating reads that cannot possibly have an adapter. 

In terms of performance on my PC the main cutadapt processes 9.4M reads per minute and with this addition reaches 16.1M reads per minute.

This still needs tests and a way to shut off the heuristic when it becomes to costly. But at least it is in a good enough state to play around with the parameters a bit and see how this can be integrated best.

</details>