@rhpvorderman 2022-12-20
While I said I didn't have time to work on this, I did most of the above work on the weekend. Yesterday I managed to benchmark it on my work machine with Intel i7-6700 processor. (My home machine is a ryzen 5 3600). Interestingly the heuristic was significantly faster on the intel machine. I checked the GNU libc source code and apparently if the processor supports fast unaligned load then it switches to a specific SSE2 version of strstr. So on this intel processor the speedup was quite significant leading to a 2x overall cutadapt speedup on the tested file from 8 microseconds to 4 microseconds per read.