@rhpvorderman 2022-12-22
I am aware that this thread is getting monologue qualities, I hope you don't mind. Just collecting my thoughts on the subject here lest they do get lost in my memory.

So strstr works quite well and has a few advantages:
- Works well out of the box
- Reasonably fast
- Fast asm implementation available. Makes it about 2x faster on intel than on amd.

It also has a few disadvantages:
- Uses C-strings, rather than memory with length. The latter is faster because strlen does not need to be called. 
- Memory with length also allows front adapters. These are now excluded.
- Every string search algorithm seems to be to consist of three fases:
  - quick heuristic to see if it can be solved with strchr or by other means.
  - initialization phase where the needle is converted in a sort of structure that allows fast checking.
  - The actual check.
  The main disadvantage here is that this initialization is performed over and over again for the same kmers. 
  
One of the disadvantages can be solved using glibc's memmem. But that is not part of the C standard, and will give headaches on windows. (Not on linux though, even musl supports it).

I was hesitant to implement something myself, but then I found this truly excellent thread on stackoverflow: https://stackoverflow.com/questions/3183582/what-is-the-fastest-substring-search-algorithm . From there was a link to this [great article that compares more than 50 string searching algorithms](https://arxiv.org/pdf/1012.2547v1.pdf). What this shows is that for the particular case of small alphabets (like DNA) and relatively short needles (like the 6, 7 and 8-mers in this algorithm) the best fit is the shift-and algorithm. The good news is that this algorithm is **very simple** there is very good C example code on wikipedia where it is known as the [bitap algorithm](https://en.wikipedia.org/wiki/Bitap_algorithm). It is about 30 lines of code. Very simple. It works by creating a lookup table of integers where each integers has zero bits in the position where the letters are. So `banana` would case the `a` entry in the lookuptable to look like this `101010 `. I benchmarked the wikipedia code and it is close to strstr.

Implementing this would have the following advantages:
- Allow the initialization outside of the find loop. So initialize each kmer once and store the lookup table. That will save a lot of time.
- Allow having memory and lengths, allowing front adapters.
- This morning I realized that the lookup table could easily be updated to also allow upper and lowercase matching. Simply make the `A` entry for `banana` also `101010`. 
- Then I realized this algorithm could therefore also easily be used to matchIUPAC characters. So the heuristic can be applied in all use cases!!! *This is awesome!* 

Just needed to post this here so the idea isn't lost. Merry christmas and a happy new year!