@rhpvorderman 2022-12-28
I implemented the Shift-OR algorithm. I erroneously thought it was shift-AND but I corrected the naming in the code when I found out how it worked exactly.  According to the paper, the performance is exactly the same between them. 
The good news is that this indeed allowed for costless matching of IUPAC bases, uppercase-lowercase etc. Also the algorithm is extremely simple and allowed separating the creation of the bitmatrix from the searchalgorithm. This makes it fast. Even faster than the SSE2 implementation of strstr. Which is great because now the cost of the heuristic seems to be 1/8 to 1/12 of an alignment depending on the machine. 
It is applicable to all sorts of adapters. A quick test on straightforward illumina data with truseq adapters:

before (main):
```
This is cutadapt 4.3.dev1+g1ae3074 with Python 3.9.2
Command line parameters: -a AGATCGGAAGAGCACACGTCTGAACTCCAGTCA /home/rhpvorderman/test/5millionreads_R1.fastq -o /dev/null
Processing single-end reads on 1 core ...
Done           00:00:41     5,000,000 reads @   8.3 µs/read;   7.19 M reads/minute
Finished in 41.724 s (8.345 µs/read; 7.19 M reads/minute).

=== Summary ===

Total reads processed:               5,000,000
Reads with adapters:                   559,384 (11.2%)
Reads written (passing filters):     5,000,000 (100.0%)

Total basepairs processed:   755,000,000 bp
Total written (filtered):    746,260,975 bp (98.8%)

=== Adapter 1 ===

Sequence: AGATCGGAAGAGCACACGTCTGAACTCCAGTCA; Type: regular 3'; Length: 33; Trimmed: 559384 times

```

after (this branch):
```
This is cutadapt 4.3.dev30+g7e3f64e with Python 3.9.2
Command line parameters: -a AGATCGGAAGAGCACACGTCTGAACTCCAGTCA /home/rhpvorderman/test/5millionreads_R1.fastq -o /dev/null
Processing single-end reads on 1 core ...
Done           00:00:19     5,000,000 reads @   3.9 µs/read;  15.31 M reads/minute
Finished in 19.596 s (3.919 µs/read; 15.31 M reads/minute).

=== Summary ===

Total reads processed:               5,000,000
Reads with adapters:                   559,384 (11.2%)
Reads written (passing filters):     5,000,000 (100.0%)

Total basepairs processed:   755,000,000 bp
Total written (filtered):    746,260,975 bp (98.8%)

=== Adapter 1 ===

Sequence: AGATCGGAAGAGCACACGTCTGAACTCCAGTCA; Type: regular 3'; Length: 33; Trimmed: 559384 times
```

Do you have a pool of real-world cutadapt test data somewhere that I can run on this branch to see if the behavior is as good for other files? Even though theoretically false negatives of the heuristic are impossible, I think extensive validation is necessary before I dare propose this change for implementation.