@rhpvorderman 2023-01-04
@marcelm. This code is ready for review (please do not haste, it is not urgent). I have added the necessary unit tests. I also verified 128 Fastq pairs using illumina adapters. The output between `main` and this branch is exactly the same. The reports are exactly the same too (except for the version, runtime and command line of course). At our institute we have lots and lots of Illumina data available so it was easy to verify that this works correctly on production data. Given that in theory it is impossible that the results deviate, and that in practice I tried over a billion reads, I think it indeed works correctly.

 For other uses of cutadapt I need a few pointers to data which I can use. We have a compute cluster so it is trivial to do a massively parallel comparison. So I don't mind doing the additional verification. 

I hope you have had a nice holiday!